backend_django/welcome_mate/requirements.txt
# Django & REST Framework
Django>=5.1.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Database
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9

# Environment
python-dotenv>=1.0.0

# AI & APIs
google-generativeai>=0.3.0
google-api-python-client>=2.100.0

# Image Processing
Pillow>=10.0.0

# Production Server
gunicorn>=21.0.0
